Make fields configurable
There's now a mapping attribute <= header field, which is stored in
the conf file. If not configured, it is initialized with a default
value.

This opens the possibility to parse previous formats with slightly
different sets of fields.

Some fields are renamed to match what is used in the CSV header and
simplify things. This was a change that we wanted to do anyway.