add a new -l option to rank.
WARNING: this changes the default ranking output:
if you do 'rank' you will get a raking based solely on people score
("fair ranking")

if you do 'rank -l' you will get a ranking that takes into account
first labels and then the score. This pushes on top of the list the
people who confirmed, then VIPs, then invited, then invited samelab,
then shortlisted. It also pushed further down those people who declined
our invitation.

This is algorithmically not a good solution. The best thing would be
to a habe Person become a full-fledge class (and not a simple
namedtuple). This would allow us to implement different __lt__, __eq__,
etc... methods, so that the sorting is concentrated in a single class
and the sorting method can be swapped at will.