Rename female to nonmale, fix stats
stat verb was broken because "Female" != "female"; use .lower() to make them
equivalent.

Also, female is replaced with nonmale in the code, and "Female" is replaced with
"Female/Other" in the output.

Our formula currently does not use this variable, but formulas for past years
would have to be adjusted.