Allow values to to be missing
Map missing programming/open source/etc values to "(none)".
This can be then used in the right sections in grader.conf to assign
a value. I think limesurvey should be configured to not allow incomplete
values like that, but if it wasn't, then let's make the best of this situation.

[open_source_rating]
never used = 0.0
...
(none) = 0

[python_rating]
competent = 1.0
...
(none) = 0

[vcs_rating]
no = 0
...
(none) = 0