Allow some reviewers to abstain
A score from a reviewer may be set to nan, in which case it is skipped
in the calculation as if the score wasn't set at all, but also the reviewer
is not asked to grade this person.

Since reviewing is blind to names, the usual way to use this would be to
pre-populate grader.conf with known people to skip, e.g.:
[motivation_score-1]
pavel  tolmachev = nan
dennis liu = nan
jianguang fang = nan

But it is also possible to mark people interactively by entering 'abstain'
in grader (or any prefix starting with 'abs').