Rework how mappings are done once again
Here the idea is that we either match a field exactly, or match a pattern
that is clearly longer than other candidates. This seems to do the right thing
for all applications lists that we have in the main repo.

I'm keeping the print statements, but disabled. They are quite useful to
debug this when necessary.