HACK! Add a way to override for person fields
This is a terrible hack. It allows us to override some of the fields in
the applications with other values. For example, if person X Y has
ranked herself as a novice in Python, but we think she is indeed
expert, we can add the following in grader.conf:

[python_overrides]
x y = expert

This override is read at the beginning when the list of applicants is
generated. No validatio whatsoever is done on the values, so beware!!!