grader/applications: when saving, write text in one step
Previously, if the generation of the ini text failed, we would end up with an
empty or truncated file. This is not very likely to happen during normal use,
but when developing the previous patch, we hit it many times. This file is very
precious, so let's make an effort to avoid that.

We now generate the contents first, and then write them using
Path.write_text(). This is not atomic, but much less likely to fail halfway.