Resolve the problem with the color prompt messing up cursor position
When scrolling through history with ↑/↓, once a sufficiently long line
was shown, the cursor would not be returned to the beginning of the line
and a portion of the previous line would be shown.

It turns out that this can be avoided if we wrap the color escape sequences
with \1 … \2, which tells the terminal (?) that the sequence is zero-width.
(The sequence is for the terminal, but maybe python, or libreadline, is
smart enough to take this into account. I'm hazy on the details.)